Meta Box back compat and fall backs (#3554)
Fall back to the classic editor when a meta box declares itself incompatible with Gutenberg, with the `__block_editor_compatible_meta_box` arg. This can later be expanded to automatically detect meta boxes that aren't compatible, and fall back.

When a meta box causes a fall back, and `WP_DEBUG` is enabled, show a warning to the end user in the meta box.