Fix: Align Hook: Impossible to set no alignment when a default align exists (#9634)
When we have a default value for an attribute, during the load of a document/block if that attribute is undefined the attribute gets the default value.
If we had a block that supports align but contained a default align we had a problem. If no align was set the post correctly saved the markup of no alignment set but the alignment attribute was not saved watching by the value of the attribute it was impossible to differentiate the state of default alignment being used, and the state of no alignment is used. The alignment attribute was just not stored in both cases.
If alignment was set to none (undefined), When the block was loaded no attribute was set, so the default value of alignment was used, but the block had the markup of alignment none, so the block becomes invalid.

This PR solves this problem by setting alignment to none when setting no alignment on the align hook, this way none is saved on the attribute and Gutenberg correctly differentiates the state of the block has the default align and the state the block has no alignment.


## How has this been tested?
I used the following test block: https://gist.github.com/jorgefilipecosta/b733f1745531dc9758c6eba0838ae3f5; I checked on the code editor that if setting no alignment for this block we set alignment comment attribute to 'none' and no align class is added. And I checked that after removing the alignment saving the post and loading it again everything works as expected. 

## Screenshots <!-- if applicable -->
Before:
![sep-05-2018 16-28-13](https://user-images.githubusercontent.com/11271197/45104639-708f9180-b12a-11e8-9ad7-bdbe34caa05e.gif)


After:
![sep-05-2018 15-46-36](https://user-images.githubusercontent.com/11271197/45104626-68cfed00-b12a-11e8-9a4d-854aa03d596e.gif)