Fix annotations text tracking (#11861)
* Allow filtering formatToValue and valueToFormat

This makes it possible to have an editor-only format that stays in
position when typing inside `RichText`.

Apply this to annotations, because it is necessary there.

* Update annotations based on changes in `RichText`

When the user types before or in an annotation we want the annotation
to be adjusted accordingly. We can determine the location of the mark
elements in the `RichText` based on the value passed to
`__experimentalCreateValueToFormat` so we can update the annotations
based on the values expressed there.

* Annotations: Refactor RichText APIs

Instead of having APIs for `formatToValue` and `valueToFormat`, create
an `__experimentalCreateOnChangeEditableValue` function that is called
when the editable value changes. The annotations API can use this to
update its state.

* Fix and add annotations reducer tests

* Move annotations file to separate file.

* Add e2e tests for annotation text tracking

* Change e2e test annotations sidebar name to be unique

* Update docs