Add: action to set category icon; Fix: Managing block categories documentation (#11640)
The documentation of "Managing block categories" was wrong as on the server we can not set SVG icons. The server would not be able to use our SVG component.
This PR fixes the documentation and adds a new action that allows updating the icon on the front end in a simple way.